requirements.txt
fastapi
uvicorn
langchain
langgraph
redis
requests
pydantic
python-dotenv
redis
folium